pyproject.toml
[tool.poetry]
name = "communautofinder-telegrambot-python"
version = "0.1.0"
description = ""
authors = ["FloDevelops <65234487+FloDevelops@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = {extras = ["webhooks"], version = "^20.6"}
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
